Only need to train the model now 1

+++ src/textSummarizer/entity/__init__.py
@@ -33,6 +33,6 @@ class ModelTrainerConfig:
     logging_steps: int
     evaluation_strategy: str
     eval_steps: int
-    save_steps: float
+    save_steps: int
     gradient_accumulation_steps: int
